Just added the logic in the PhysicsEngine for a planetary type motion to be induced on the bodies once they are added into the space.
Ended up being able to remove the player class, I can definietly see the
removal of the 'movable' module coming very soon as well. Just pruning
anything that doesn't need to happen in order to have a working game.

Such as: the Physics engine is only talking to PolyBounce, so there no
need to include a observer system. It;s too verbose for this
application.

PolyBounce is just going to pull the updates as needed from the
PhysicsEngine.space. Much cleaner approach.

May need to create a quick timer class for the rings.

Haven't touched the rings at all yet. I know it's going to be complex,
so I'm wokring my way up there.

Everything is working perfectly so far. Making sure that I'm running the
game every time I make a change in case there's anythig that needs to be
addressed, it's happening before it becomes to complex to solve the
issue or unravel the layers upon layers of abstraction I've already
implemented depending on faulty logic, etc.

Exciting stuff. I should've come to the library much sooner than this.
Buts OK. Gonna be great.